# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST helper endpoints for AMG X-wing squads",
    "version": "2.5.0"
  },
  "host": "squad2xws.herokuapp.com",
  "basePath": "",
  "schemes": ["http"],
  "produces": ["application/json"],
  "paths": {
    "/yasb/xws": {
      "get": {
        "tags": ["XWS"],
        "summary": "Translates YASB url to XWS output",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": [
          {
            "name": "f",
            "in": "query",
            "description": "faction (ex: 'Galactic Empire')",
            "required": true,
            "type": "string",
            "enum": ["Rebel Alliance", "Galactic Empire", "Scum and Villainy", "First Order", "Resistance", "Galactic Republic", "Separatist Alliance"]
          },
          {
            "name": "d",
            "in": "query",
            "description": "serialized squad (example 'v8ZsZ200Z179X116W249W165Y166X112WWWWW162Y191X240W136WWWW71WW')",
            "required": true,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "query",
            "description": "squad name (ex: 'My Awesome Squad')",
            "required": false,
            "type": "string"
          },
          {
            "name": "obs",
            "in": "query",
            "description": "obstacles list, comma separated (ex: 'coreasteroid5,core2asteroid4,core2asteroid5')",
            "required": false,
            "type": "string"
          }
        ]
      }
    },
    "/yasb/reverse": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["XWS"],
        "summary": "Translates XWS input to YASB URL",
        "responses": {
          "200": {
            "description": "YASB URL string",
            "schema": {
              "type": "object",
              "example": {
                "url": "https://raithos.github.io/?f=Galactic%20Empire&d=v8ZsZ200Z179X116W165W249Y191X240W136WWWW71WWY167X112WW28WWW&sn=Soontir%20Jenden%20Redline&obs="
              },
              "application/json": {
                "example": {
                  "url": "https://raithos.github.io/?f=Galactic%20Empire&d=v8ZsZ200Z179X116W165W249Y191X240W136WWWW71WWY167X112WW28WWW&sn=Soontir%20Jenden%20Redline&obs="
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "requestBody": {
          "description": "XWS JSON representing x-wing squad",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prop": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "example": {"faction":"galacticempire","pilots":[{"id":"soontirfel","ship":"tieininterceptor","upgrades":{"modification":["shieldupgrade","targetingcomputer"],"ept":["crackshot"]},"points":65},{"id":"redline","ship":"tiecapunisher","upgrades":{"torpedo":["protontorpedoes"],"sensor":["passivesensors"],"device":["seismiccharges"]},"points":71},{"id":"coloneljendon","ship":"lambdaclasst4ashuttle","upgrades":{"crew":["directorkrennic"],"sensor":["collisiondetector"]},"points":56}],"name":"Soontir Jenden Redline","description":"","points":192,"vendor":{"sb2xws":{"builder":"SB2XWS","url":"http://sb2xws.herokuapp.com/translate/c230c0fd-9461-41c9-9a39-e0f9b641c094"}}}
        }
      }
    }
  }
}